--- addons/mind_game/scenes/SavedConversation.tscn
@@ -1,21 +1,0 @@
-[gd_scene load_steps=2 format=3 uid="uid://b84iwhex1h0xb"]
-
-[ext_resource type="Texture2D" uid="uid://5ga0x3yawju4" path="res://addons/mind_game/assets/game_icons/PNG/White/1x/trashcan.png" id="1_bufxj"]
-
-[node name="SavedConversationHBoxContainer" type="HBoxContainer"]
-offset_right = 290.0
-offset_bottom = 58.0
-size_flags_horizontal = 3
-size_flags_vertical = 0
-
-[node name="ConversationLoadButton" type="Button" parent="."]
-unique_name_in_owner = true
-layout_mode = 2
-size_flags_horizontal = 3
-text = "Saved Conversation"
-
-[node name="ConversationDeleteButton" type="Button" parent="."]
-layout_mode = 2
-size_flags_vertical = 0
-icon = ExtResource("1_bufxj")
-icon_alignment = 1
