+++ addons/mind_game/scripts/model_classes/MindAgent.cs
@@ -6,7 +6,8 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace MindGame
-{ 
+{
+    [Tool]
     public partial class MindAgent : Node
     {
