+++ project.godot
@@ -11,7 +11,7 @@ config_version=5
 [application]
 
 config/name="MindGame"
-run/main_scene="res://addons/mind_game/examples/single_agent_chat/SingleAgentChat.tscn"
+run/main_scene="res://addons/mind_game/controllers/SingleAgentChatController.tscn"
 config/features=PackedStringArray("4.3", "C#", "Mobile")
 config/icon="res://addons/mind_game/assets/logos/mind_game_logo_01.png"
